feat: ListActivity 서버 통신 구현

+++ app/src/main/java/com/sieun/supporters_android/api/ApiService.kt
@@ -11,5 +11,5 @@ interface ApiService {
     suspend fun getHome(): Response<HomeResult>
 
     @GET("/sponsors/categories/{categoryNum}/items")
-    suspend fun fetchCategoryItem(@Path("categoryNum") categoryNum: Int): Response<CategoryItemsResult>
+    suspend fun fetchCategoryItem(@Path("categoryNum") categoryNum: Long): Response<CategoryItemsResult>
 }
